Add support for scalar subqueries in delete queries
